requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.2
Faker==18.3.0
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.0.3
pydantic_core==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
